+++ changelog.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 9.12.0
+
+* `[knf]` Added new validators: `NotLen`, `NotPrefix` and `NotSuffix`
+* `[knf]` Validators code refactoring
+
 ### 9.11.2
 
 * `[system/process]` Fixed bug with parsing CPU data